refactor mocp control to separate file
