+++ src/lib.rs
@@ -22,6 +22,7 @@ pub mod input;
 pub mod keybindings;
 pub mod log;
 pub mod marks;
+pub mod mocp;
 pub mod mode;
 pub mod opener;
 pub mod preview;
